refactor(mesh): rename components
