+++ packages/mesh/app/media/_components/publisher-card.tsx
